src/test/java/com/example/web/CustomerServiceTest.java
package com.example.web;

import com.example.entity.domain.Customer;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

@RunWith(SpringRunner.class)
@SpringBootTest
public class CustomerServiceTest {

    @Autowired
    private CustomerService customerService;

    @Test
    public void sucess_getCustomer(){

        List<Customer> actualCustomerList = customerService.getCustomer();

        assertThat(actualCustomerList).isNotNull();
    }
}
